* Removed warnings.

+++ src/pack.c
@@ -60,8 +60,8 @@
 #define htonll(x)   (x)
 #define ntohll(x)   (x)
 #else /* HAS_BIGENDIAN */
-#define htonll(x)   ((((uint64_t)htonl(x)) << 32) + htonl(x >> 32))
-#define ntohll(x)   ((((uint64_t)ntohl(x)) << 32) + ntohl(x >> 32))
+#define htonll(x)   ((((uint64_t)htonl((uint32_t)x))<<32) + htonl((uint32_t)(x>>32)))
+#define ntohll(x)   ((((uint64_t)ntohl((uint32_t)x))<<32) + ntohl((uint32_t)(x>>32)))
 #endif /* HAS_BIGENDIAN */
